Major HRM and thumb refactor.
HRM moved one key inward since we don't use the shift very often on the
index. This allows us to keep the pinky free of any holds (except the fn
layer).

Space is not a hold tap anymore because when typing quickly we often
trigger the hold function.

Backspace and tab are the outermost thumb keys.
Shift+Backspace is now the delete key.
Minus and Sqt are back on the extra outer keys. This seems more natural
than the thumb keys where we kept trying to hit it with one of the
fingers.

The mac layers use more transparent keys to make the definition shorter.